backend_django/board_1/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
autopep8==1.5.6
black==22.3.0
blinker==1.5
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
collection==0.1.6
colorama==0.4.4
crypto==1.4.1
cryptography==37.0.4
cycler==0.10.0
distlib==0.3.4
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
filelock==3.7.0
gunicorn==20.1.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
kaitaistruct==0.10
kiwisolver==1.3.2
matplotlib==3.4.3
mpmath==1.2.1
mypy-extensions==0.4.3
Naked==0.1.31
npm==0.1.1
numpy==1.20.2
optional-django==0.1.0
outcome==1.2.0
pandas==1.3.3
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.21
pycryptodome==3.14.1
pycryptodomex==3.11.0
pydivert==2.1.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pythonping==1.1.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
selenium==4.3.0
selenium-wire==4.6.5
shellescape==3.8.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
sympy==1.9
toml==0.10.2
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.2.0
tzdata==2022.2
urllib3==1.26.9
virtualenv==20.14.1
whitenoise==6.2.0
wsproto==1.1.0
zstandard==0.18.0
